chore(crypto): Renamed files to single syntax
